requirements.txt
redis==6.0.0
dj-rest-auth==5.0.1
django-cors-headers==4.3.0
python-dotenv==1.0.1
Django==5.2
django-allauth==0.58.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
psycopg2-binary
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
pillow==11.2.1
weasyprint==65.1
cloudinary==1.44.0
python-dotenv==1.0.1
setuptools==70.0.0